--- src/main/java/com/GameStartMain.java
@@ -1,10 +1,0 @@
-package com;
-
-import com.view.InputView;
-
-public class GameStartMain {
-    public static void main(String[] args) {
-        NumberSet inputNumbers = new InputView().input();
-        new GameProcessor().start(inputNumbers);
-    }
-}
