--- src/main/java/com/view/ResultView.java
@@ -1,14 +1,0 @@
-package com.view;
-
-import com.Car;
-
-import java.util.List;
-
-public class ResultView {
-    public void showResult(List<Car> list) {
-        for (Car car : list) {
-            System.out.println(car.getResult());
-        }
-        System.out.println();
-    }
-}
